requirements.txt
selenium
pyyaml
requests
pytest
seleniumbase